+++ .idea/misc.xml
@@ -14,7 +14,7 @@
         <entry key="app/src/main/res/drawable/ic_github_outlined.xml" value="0.1705" />
         <entry key="app/src/main/res/drawable/ic_search.xml" value="0.1705" />
         <entry key="app/src/main/res/layout/activity_main.xml" value="0.28958333333333336" />
-        <entry key="app/src/main/res/layout/fragment_home.xml" value="0.3029891304347826" />
+        <entry key="app/src/main/res/layout/fragment_home.xml" value="0.31160896130346233" />
         <entry key="app/src/main/res/layout/load_state_layout.xml" value="0.3029891304347826" />
         <entry key="app/src/main/res/layout/repository_item_layout.xml" value="0.2796610169491525" />
       </map>
