+++ app/src/main/java/ge/example/githubapidemo/feature_github_repositories/domain/use_cases/SearchRepositoryUseCase.kt
@@ -8,7 +8,7 @@ import kotlinx.coroutines.flow.Flow
 class SearchRepositoryUseCase(private val githubRepository: GithubRepository) {
     suspend operator fun invoke(
         query: String,
-        perPage: Int = 30,
+        perPage: Int = 20,
         page: Int = 1
     ): Resource<GithubResponse> {
         return githubRepository.searchRepository(query, perPage, page)
