+++ app/src/main/java/ge/example/githubapidemo/feature_github_repositories/presentation/fragment/home/adapter/RepositoryAdapter.kt
@@ -6,19 +6,19 @@ import android.view.ViewGroup
 import androidx.paging.PagingDataAdapter
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.RecyclerView
+import ge.example.githubapidemo.R
 import ge.example.githubapidemo.databinding.RepositoryItemLayoutBinding
-import ge.example.githubapidemo.extensions.setNetworkImage
-import ge.example.githubapidemo.feature_github_repositories.domain.model.RepositoryItem
+import ge.example.githubapidemo.feature_github_repositories.domain.model.GithubResponse
 
 class RepositoryAdapter :
-    PagingDataAdapter<RepositoryItem, RepositoryAdapter.ShortItemViewHolder>(REPOSITORY_COMPARATOR) {
+    PagingDataAdapter<GithubResponse.RepositoryItem, RepositoryAdapter.ShortItemViewHolder>(REPOSITORY_COMPARATOR) {
 
     companion object {
-        private val REPOSITORY_COMPARATOR = object : DiffUtil.ItemCallback<RepositoryItem>() {
-            override fun areItemsTheSame(oldItem: RepositoryItem, newItem: RepositoryItem) =
+        private val REPOSITORY_COMPARATOR = object : DiffUtil.ItemCallback<GithubResponse.RepositoryItem>() {
+            override fun areItemsTheSame(oldItem: GithubResponse.RepositoryItem, newItem: GithubResponse.RepositoryItem) =
                 oldItem.id == newItem.id
 
-            override fun areContentsTheSame(oldItem: RepositoryItem, newItem: RepositoryItem) =
+            override fun areContentsTheSame(oldItem: GithubResponse.RepositoryItem, newItem: GithubResponse.RepositoryItem) =
                 oldItem == newItem
         }
     }
@@ -40,13 +40,14 @@ class RepositoryAdapter :
     inner class ShortItemViewHolder(private val binding: RepositoryItemLayoutBinding) :
         RecyclerView.ViewHolder(binding.root), View.OnClickListener {
 
-        fun onBind(repo: RepositoryItem?) {
+        fun onBind(repo: GithubResponse.RepositoryItem?) {
             repo?.apply {
                 binding.apply {
 //                    ownerImage.setNetworkImage(owner?.avatarUrl)
                     repoNameTv.text = fullName
-                    repoDescTv.text = description
-                    languageTv.text = language ?: "None"
+                    repoDescTv.text =
+                        description ?: repoDescTv.context.getString(R.string.no_repo_desc)
+                    languageTv.text = language ?: languageTv.context.getString(R.string.none)
                     startTv.text = stargazersCount.toString()
                     forksTv.text = forksCount.toString()
                 }
