+++ app/src/main/res/layout/activity_main.xml
@@ -10,6 +10,7 @@
         android:id="@+id/main_toolbar"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
+        android:background="?attr/colorPrimary"
         android:paddingTop="10dp"
         android:paddingBottom="10dp"
         app:contentInsetStartWithNavigation="0dp"
