+++ app/src/main/res/layout/repository_item_layout.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
     android:layout_margin="10dp"
-    app:cardBackgroundColor="?attr/colorPrimary"
+    app:cardBackgroundColor="?attr/colorOnSurface"
     app:cardCornerRadius="10dp"
     app:cardElevation="0dp"
     app:strokeWidth="0dp">
@@ -13,7 +13,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:background="?attr/colorPrimary"
+        android:background="?attr/colorOnSurface"
         android:orientation="vertical"
         android:padding="15dp">
