+++ app/src/main/res/values-night/themes.xml
@@ -2,16 +2,16 @@
     <!-- Base application theme. -->
     <style name="Theme.GithubApiDemo" parent="Theme.MaterialComponents.DayNight.NoActionBar">
         <!-- Primary brand color. -->
-        <item name="colorPrimary">@color/primaryDarkVariant</item>
-        <item name="colorPrimaryVariant">@color/primaryDark</item>
+        <item name="colorPrimary">@color/primaryDark</item>
+        <item name="colorPrimaryVariant">@color/primaryDarkVariant</item>
         <item name="colorOnPrimary">@color/black</item>
+        <item name="colorOnSurface">@color/primaryDarkVariant</item>
         <!-- Secondary brand color. -->
         <item name="colorSecondary">@color/primaryVariant</item>
         <item name="colorSecondaryVariant">@color/primary</item>
         <item name="colorOnSecondary">@color/white</item>
         <!-- Status bar color. -->
-        <item name="android:statusBarColor" tools:targetApi="l">?attr/colorPrimaryVariant</item>
+        <item name="android:statusBarColor" tools:targetApi="l">?attr/colorPrimary</item>
         <!-- Customize your theme here. -->
-        <item name="android:background">@color/primaryDark</item>
     </style>
 </resources>
