+++ app/src/main/res/values/strings.xml
@@ -4,4 +4,6 @@
     <string name="hello_blank_fragment">Hello blank fragment</string>
     <string name="your_feed">Your <b>Feed</b></string>
     <string name="beedep_forked_from_bprayush_devstack_from_edx_devstack"><b>beedep </b> forked from <b> bprayush/devstack</b> from <b>edx/devstack</b> </string>
+    <string name="no_repo_desc">Repo doesn\'t have any description</string>
+    <string name="none">None</string>
 </resources>
